backend/requirements.txt
google-cloud-videointelligence
pymongo
flask
opencv-python
